refactor: 型のエラーを修正する
