+++ system/index.ts
@@ -1,12 +1,13 @@
 import { factory } from "~/system/factory"
+
 import * as index from "~/system/routes/index"
-import * as message$roomId from "~/system/routes/message.$roomId"
-import * as messages from "~/system/routes/messages"
+import * as projects from "~/system/routes/projects"
+import * as projects$project from "~/system/routes/projects.$project"
 
 export const hono = factory
   .createApp()
   .post("/", ...index.POST)
-  .get("/messages", ...messages.GET)
-  .get("/message/:roomId", ...message$roomId.GET)
+  .get("/projects", ...projects.GET)
+  .get("/projects/:project/messages", ...projects$project.GET)
 
 export type AppType = typeof hono
