+++ types/workspace.ts
@@ -1,0 +1,6 @@
+import type { Message } from "ai"
+
+export type Project = {
+  id: string
+  messages: Message[]
+}
